Split groundwater_depletion.rs into subdirectory modules
Break the 701-line monolithic file into focused sub-modules:
- types.rs: constants and GroundwaterDepletionState resource (96 lines)
- systems.rs: helper functions, update system, and plugin (160 lines)
- tests.rs: all unit tests (442 lines)
- mod.rs: re-exports everything publicly for API compatibility

Closes #1389
